# services/control-plane/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DecisionDetail": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SafeCanonicalAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Type"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "audit": {
            "$ref": "#/components/schemas/AuditContinuity"
          },
          "behavioral_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Behavioral Score",
            "type": "number"
          },
          "blast_radius": {
            "title": "Blast Radius",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detail_available": {
            "title": "Detail Available",
            "type": "boolean"
          },
          "detail_deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail Deleted At"
          },
          "evidence_degraded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Degraded"
          },
          "evidence_state": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence State"
          },
          "input_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Input Digest",
            "type": "string"
          },
          "mitre_mappings": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Mappings"
          },
          "pipeline_timings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Timings"
          },
          "policy_decision": {
            "enum": [
              "ALLOW",
              "DENY",
              "REQUIRE_APPROVAL",
              "NOT_APPLICABLE"
            ],
            "title": "Policy Decision",
            "type": "string"
          },
          "policy_matches": {
            "items": {
              "$ref": "#/components/schemas/PolicyMatchDetail"
            },
            "title": "Policy Matches",
            "type": "array"
          },
          "policy_pack_version": {
            "title": "Policy Pack Version",
            "type": "string"
          },
          "predicted_next": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Next"
          },
          "reasons": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasons"
          },
          "response_status": {
            "enum": [
              "PENDING",
              "DELIVERY_PENDING",
              "DELIVERED",
              "FAILED",
              "EXECUTED"
            ],
            "title": "Response Status",
            "type": "string"
          },
          "similar_decisions": {
            "items": {
              "$ref": "#/components/schemas/SimilarDecisionReference"
            },
            "title": "Similar Decisions",
            "type": "array"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "ALLOW",
              "WOULD-BLOCK",
              "BLOCK"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "created_at",
          "agent_id",
          "action_type",
          "verdict",
          "behavioral_score",
          "blast_radius",
          "policy_decision",
          "evidence_degraded",
          "response_status",
          "detail_available",
          "input_digest",
          "action",
          "reasons",
          "mitre_mappings",
          "predicted_next",
          "policy_pack_version",
          "policy_matches",
          "evidence_state",
          "pipeline_timings",
          "audit",
          "similar_decisions",
          "detail_deleted_at"
        ],
        "title": "DecisionDetail",
        "type": "object"
      },
      "DecisionPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DecisionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "snapshot_at": {
            "format": "date-time",
            "title": "Snapshot At",
            "type": "string"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "snapshot_at"
        ],
        "title": "DecisionPage",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "EvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/CanonicalAction"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceEnvelope"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "action",
          "evidence",
          "idempotency_key"
        ],
        "title": "EvaluationRequest",
        "type": "object"
      },
      "EvaluationResponse": {
        "additionalProperties": false,
        "properties": {
          "behavioral_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Behavioral Score",
            "type": "number"
          },
          "blast_radius": {
            "title": "Blast Radius",
            "type": "string"
          },
          "evidence_degraded": {
            "title": "Evidence Degraded",
            "type": "boolean"
          },
          "matched_rules": {
            "items": {
              "$ref": "#/components/schemas/PolicyMatchResponse"
            },
            "title": "Matched Rules",
            "type": "array"
          },
          "mitre_atlas": {
            "items": {
              "type": "string"
            },
            "title": "Mitre Atlas",
            "type": "array"
          },
          "mitre_attack": {
            "items": {
              "type": "string"
            },
            "title": "Mitre Attack",
            "type": "array"
          },
          "pipeline_timings": {
            "$ref": "#/components/schemas/PipelineTimings"
          },
          "policy_decision": {
            "enum": [
              "ALLOW",
              "DENY",
              "REQUIRE_APPROVAL"
            ],
            "title": "Policy Decision",
            "type": "string"
          },
          "policy_pack_version": {
            "title": "Policy Pack Version",
            "type": "string"
          },
          "predicted_next": {
            "title": "Predicted Next",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "response_status": {
            "enum": [
              "DECIDED",
              "DELIVERY_PENDING"
            ],
            "title": "Response Status",
            "type": "string"
          },
          "similar_decision_ids": {
            "items": {
              "type": "string"
            },
            "title": "Similar Decision Ids",
            "type": "array"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "ALLOW",
              "WOULD-BLOCK",
              "BLOCK"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "verdict",
          "behavioral_score",
          "blast_radius",
          "reasons",
          "reason_codes",
          "mitre_attack",
          "mitre_atlas",
          "predicted_next",
          "policy_decision",
          "policy_pack_version",
          "matched_rules",
          "evidence_degraded",
          "response_status",
          "pipeline_timings",
          "similar_decision_ids"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyMatchDetail": {
        "additionalProperties": false,
        "properties": {
          "effect": {
            "enum": [
              "ALLOW",
              "DENY",
              "REQUIRE_APPROVAL"
            ],
            "title": "Effect",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "rule_version": {
            "title": "Rule Version",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "rule_version",
          "effect",
          "metadata"
        ],
        "title": "PolicyMatchDetail",
        "type": "object"
      },
...
      "SafeCanonicalAction": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "consequential": {
            "title": "Consequential",
            "type": "boolean"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "action_type",
          "target",
          "consequential",
          "attributes"
        ],
        "title": "SafeCanonicalAction",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A separately deployed, multi-tenant service. The legacy Flask /v1/gate boundary is not part of this application.",
    "summary": "Production security decision control plane",
    "title": "DUSK Control Plane API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/decisions": {
      "get": {
        "operationId": "list_decisions_v2_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "in": "query",
            "name": "verdict",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ALLOW",
                    "WOULD-BLOCK",
                    "BLOCK"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "in": "query",
            "name": "policy_decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ALLOW",
                    "DENY",
                    "REQUIRE_APPROVAL",
                    "NOT_APPLICABLE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Policy Decision"
            }
          },
          {
            "in": "query",
            "name": "response_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "DELIVERY_PENDING",
                    "DELIVERED",
                    "FAILED",
                    "EXECUTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Response Status"
            }
          },
          {
            "in": "query",
            "name": "evidence_degraded",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evidence Degraded"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "in": "query",
            "name": "action_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action Type"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Decisions",
        "tags": [
          "decisions"
        ]
      }
    },
    "/v2/decisions/{trace_id}": {
      "get": {
        "operationId": "get_decision_v2_decisions__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Decision",
        "tags": [
          "decisions"
        ]
      }
    },
    "/v2/evaluations": {
      "post": {
        "operationId": "evaluate_action_v2_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evaluate Action",
        "tags": [
          "evaluations"
        ]
      }
    }
  }
}